Travail sur la navigation secondaire en mobile
